Create a stockpicker
This program iterates over the stock array two times, comparing a value
from the array to all the values after that compared one, and a profit
is calculated. If the profit is over the one of a previously calculated
one, it overwrite the previous day pair for which stock is better to
buy and sell.